[New Feature] Code refactor DistriOptimizer for advance parameter operations  (#2556)

+++ scala/dllib/src/main/scala/com/intel/analytics/bigdl/dllib/tensor/SparseTensor.scala
@@ -1121,7 +1121,7 @@ private[tensor] class SparseTensor[@specialized(Float, Double) T: ClassTag](
     throw new UnsupportedOperationException(s"SparseTensor: Unimplemented method")
   }
 
-  override def clamp(minValue: Float, maxValue: Float): Tensor[T] =
+  override def clamp(minValue: Double, maxValue: Double): Tensor[T] =
     throw new UnsupportedOperationException(s"SparseTensor: Unimplemented method")
 
   override def sumSquare(): T =
